Refactor achievement and authentication modules to include user context. Remove calls to AppRemoteDatabase in the app.

+++ composeApp/src/commonMain/kotlin/cz/cvut/docta/achievement/data/repository/AchievementRemoteRepository.kt
@@ -10,8 +10,9 @@ class AchievementRemoteRepository: AchievementRepository {
         return emptyList()
     }
 
-    override suspend fun getUserAchievements(userId: Long): List<UserAchievementDto> {
+    override suspend fun getUserAchievements(userId: Int): List<UserAchievementDto> {
         // TODO-ACHIEVEMENTS: Implement fetching user achievements from the server
         return emptyList()
     }
+
 }
